refactor(exec) rename package 'nw' to 'ex'
