+++ src/ex/apiExec.go
@@ -1,4 +1,4 @@
-package nw
+package ex
 
 import (
     "net/http"
