+++ src/ex/intervalExec.go
@@ -1,4 +1,4 @@
-package nw
+package ex
 import(
     "time"
     "sync"
